pyproject.toml
[tool.poetry]
name = "repl_python3_musebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.2"
youtube_dl = "^2021.6.6"
google-api-python-client = "^2.33.0"
"discord.py" = "^1.7.3"
Flask = "^2.0.3"
discord-pretty-help = "^1.3.3"
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
